# todo-server/src/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "TODO REST_API",
    "description": "API para gerenciar tarefas"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/todos/{uid}": {
      "get": {
        "tags": [
          "Todo"
        ],
        "description": "Endpoint para obter todas as tarefas",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Todo"
        ],
        "description": "Endpoint para criar uma tarefa",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do usuário"
          },
          {
            "name": "data",
            "description": "Dados da tarefa",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Todo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/todos/todo/{uid}/{id}": {
      "get": {
        "tags": [
          "Todo"
        ],
        "description": "Endpoint para obter uma tarefa",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/todos/{uid}/{id}": {
      "put": {
        "tags": [
          "Todo"
        ],
        "description": "Endpoint para atualizar uma tarefa",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "done": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Todo"
        ],
        "description": "Endpoint para remover uma tarefa",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Todo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "5f9f5b9b9b9b9b9b9b9b9b9b"
        },
        "title": {
          "type": "string",
          "example": "Estudar"
        },
        "description": {
          "type": "string",
          "example": "Estudar NodeJS"
        },
        "done": {
          "type": "boolean",
          "example": false
        }
      }
    },
...
  }
}